Preparing infrastructure to adopt packed trace

+++ proofman/src/recursion.rs
@@ -32,8 +32,8 @@ pub struct MaxSizes {
     pub aux_trace_area: u64,
     pub aux_trace_recursive_area: u64,
     pub total_const_area_aggregation: u64,
-    pub n_streams: usize,
-    pub n_recursive_streams: usize,
+    pub n_streams: u64,
+    pub n_recursive_streams: u64,
 }
 
 #[derive(Debug, BorshSerialize, BorshDeserialize)]
